move SystemType to separate file, to be used in client as well
