+++ packages/client/src/v2-events/features/workqueues/EventOverview/EventOverview.stories.tsx
@@ -19,7 +19,7 @@ import {
   ActionStatus,
   getUUID
 } from '@opencrvs/commons/client'
-import { SystemType } from '@opencrvs/commons'
+import { SystemType } from '@opencrvs/commons/client'
 import { AppRouter, TRPCProvider } from '@client/v2-events/trpc'
 import { ROUTES, routesConfig } from '@client/v2-events/routes'
 import { tennisClubMembershipEventDocument } from '@client/v2-events/features/events/fixtures'
