+++ packages/client/src/views/SysAdmin/Config/Systems/useSystems.ts
@@ -25,12 +25,11 @@ import {
   RegisterSystemMutation,
   RegisterSystemMutationVariables,
   System,
-  SystemType,
   UpdatePermissionsMutation,
   UpdatePermissionsMutationVariables,
   WebhookPermission
 } from '@client/utils/gateway'
-
+import { SystemType } from '@opencrvs/commons/client'
 import React, { useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import * as mutations from './mutations'
