+++ packages/commons/src/index.ts
@@ -17,6 +17,7 @@ export * from './roles'
 export * from './search'
 export * from './events'
 export * from './users/service'
+export * from './users/systemType'
 export * from './authentication'
 export * from './utils'
 export * from './countryconfig'
