+++ packages/commons/src/users/systemType.ts
@@ -8,16 +8,10 @@
  *
  * Copyright (C) The OpenCRVS Authors located at https://github.com/opencrvs/opencrvs-core/blob/master/AUTHORS.
  */
-export * from './search'
-export * from './events'
-export * from './scopes'
-export * from './documents'
-export * from './workqueues'
-export * from './uuid'
-export * from './utils'
-export * from './conditionals/validate'
-export { DEFAULT_ROLES_DEFINITION } from './authentication'
-export * from './fixtures'
-export * from './users/systemType'
-
-export type PartialBy<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>
+export enum SystemType {
+  Health = 'HEALTH',
+  NationalId = 'NATIONAL_ID',
+  RecordSearch = 'RECORD_SEARCH',
+  Import = 'IMPORT',
+  Webhook = 'WEBHOOK'
+}
